+++ PSLedger/PSLedger.psm1
@@ -8,6 +8,7 @@
 . $PSScriptRoot\Private\ExtensionLoader.ps1
 . $PSScriptRoot\Private\ResolveJournalPath.ps1
 . $PSScriptRoot\Private\ResolveFiscalYear.ps1
+. $PSScriptRoot\Private\PathSafety.ps1
 . $PSScriptRoot\Private\OpeningBalance.ps1
 . $PSScriptRoot\Private\JournalSchema.ps1
 . $PSScriptRoot\Private\JournalMetadata.ps1